+++ microprojects/__main__.py
@@ -1,5 +1,12 @@
 def main():
-    print("Did you mean: python -m microprojects.calc")
+    """Running microprojects as a module causes ambuiguity because
+    we provides multiple executable scripts.
+
+    Did you mean:
+        python -m microprojects.calc
+        python -m microprojects.ngit
+    """
+    print(main.__doc__)
 
 
 if __name__ == "__main__":